+++ contracts/VoucherHub.sol
@@ -28,6 +28,7 @@ contract VoucherHub is
     // Granted to msg.sender == defaultAdmin() == owner()
     // DEFAULT_ADMIN_ROLE
 
+    // TODO: Rename roles.
     // Upgrade VoucherHub and Vouchers contracts.
     // Granted to msg.sender
     bytes32 public constant UPGRADE_MANAGER_ROLE = keccak256("UPGRADE_MANAGER_ROLE");
@@ -166,16 +167,32 @@ contract VoucherHub is
         uint256 value
     ) external onlyRole(VOUCHER_MANAGER_ROLE) returns (address voucherAddress) {
         VoucherHubStorage storage $ = _getVoucherHubStorage();
-        uint256 voucherExpiration = block.timestamp + getVoucherType(voucherType).duration;
+        uint256 expiration = block.timestamp + getVoucherType(voucherType).duration;
         voucherAddress = address(new VoucherProxy{salt: _getCreate2Salt(owner)}($._voucherBeacon));
         // Initialize the created proxy contract.
         // The proxy contract does a delegatecall to its implementation.
         // Re-Entrancy safe because the target contract is controlled.
-        Voucher(voucherAddress).initialize(owner, address(this), voucherExpiration, voucherType);
+        Voucher(voucherAddress).initialize(owner, address(this), expiration, voucherType);
         IERC20($._iexecPoco).transfer(voucherAddress, value); // SRLC
         _mint(voucherAddress, value); // VCHR
         $._isVoucher[voucherAddress] = true;
-        emit VoucherCreated(voucherAddress, owner, voucherExpiration, voucherType, value);
+        emit VoucherCreated(voucherAddress, owner, voucherType, expiration, value);
+    }
+
+    /**
+     * Top up a voucher by increasing its balance and pushing its expiration date.
+     * @param voucher The address of the voucher.
+     * @param value The amount of credits to top up.
+     */
+    function topUpVoucher(address voucher, uint256 value) external onlyRole(VOUCHER_MANAGER_ROLE) {
+        require(value > 0, "VoucherHub: no value");
+        VoucherHubStorage storage $ = _getVoucherHubStorage();
+        require($._isVoucher[voucher], "VoucherHub: unknown voucher");
+        _mint(voucher, value); // VCHR
+        IERC20($._iexecPoco).transfer(voucher, value); // SRLC
+        uint256 expiration = block.timestamp + $.voucherTypes[Voucher(voucher).getType()].duration;
+        Voucher(voucher).setExpiration(expiration);
+        emit VoucherToppedUp(voucher, expiration, value);
     }
 
     /**